Earthquake prediction model
This model takes three inputs: the time since the last earthquake, the intensity of the last earthquake, and the current level of fault activity. These inputs are combined using a set of coefficients, which are used to make the prediction. The prediction is then compared to a set of thresholds to determine the likelihood of an earthquake occurring.